Move text field controller outside of components directory
